requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
django-paypal==2.0
idna==3.4
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
